fix: Move the _NginxConf parser out of the combiner
* Due to an issue found with the httpdconf parser that was in
  the combiner I'm moving the nginx parser out of the combiner as well.
* Moved a parsr parser that was shared between two different parsers to
  under the parsr directory.
* Added docstring to the moved parser, and added doctest to both
  parsers.
