+++ insights/parsers/tests/test_nginx_conf.py
@@ -1,3 +1,5 @@
+import doctest
+
 from insights.parsers import nginx_conf
 from insights.tests import context_wrap
 
@@ -113,3 +115,34 @@ def test_nginxconfiguration():
 '$status $body_bytes_sent "$http_referer" '
 '"$http_user_agent" "$http_x_forwarded_for"'"""
     assert nginxconf['http']['server'][2]['location'][0]['location'][0]['limit_except']['allow'] == '192.168.2.0/32'
+
+
+def test_nginxconfpeg():
+    nginxconf = nginx_conf.NginxConfPEG(context_wrap(NGINXCONF))
+    assert nginxconf['user'][-1].value == 'root'
+    assert nginxconf['events'][-1]['worker_connections'][-1].value == 4096
+    assert nginxconf['mail'][-1]['server'][0]['listen'][-1].value == 143
+    assert nginxconf['http'][-1]['access_log'][-1].value == 'logs/access.log main'
+    assert nginxconf['http'][-1]['server'][0]['location'][0]['fastcgi_pass'][-1].value == '127.0.0.1:1025'
+    assert nginxconf['http'][-1]['server'][1]['location'][-1].value == '/'
+    assert nginxconf['http'][-1]['upstream'][1].value == 'big_server_com'
+    assert nginxconf["http"][-1]["include"][0].value == 'conf/mime.types'
+    assert nginxconf['http'][-1]['upstream'][1]['server'][0].value == '127.0.0.3:8000 weight=5'
+    assert nginxconf['http'][-1]['log_format'][
+               -1].value == 'main $remote_addr - $remote_user [$time_local] "$request"  $status $body_bytes_sent "$http_referer"  "$http_user_agent" "$http_x_forwarded_for"'
+    assert nginxconf['http'][-1]['server'][2]['location'][0]['location'][0]['limit_except'][-1]['allow'][
+               -1].value == '192.168.2.0/32'
+    assert nginxconf['http']['server']['location']['location']['limit_except']['allow'][-1].value == '192.168.2.0/32'
+    assert nginxconf['http']['server'][0]['location'][-1].value == r'~ \.php$'
+    assert nginxconf['http']['server'][1]['location'][0].value == '~ ^/(images|javascript|js|css|flash|media|static)/'
+    assert nginxconf['http']['server'][1]['location'][-1].value == '/'
+    assert nginxconf['http']['server'][-1] == nginxconf['http']['server'][2]
+
+
+def test_doc():
+    env = {
+        'nginxconf': nginx_conf.NginxConf(context_wrap(NGINXCONF)),
+        'nginxconfpeg': nginx_conf.NginxConfPEG(context_wrap(NGINXCONF)),
+    }
+    failed, total = doctest.testmod(nginx_conf, globs=env)
+    assert failed == 0
